+++ BackEnd/Lesson_10_2023_05_25/src/main/java/de/aitTr/App/CityService.java
@@ -7,7 +7,7 @@ public interface CityService {
     List<City> getAllCities();
     List<City> getAllCities(String country);
 
-    City getCity(int id);
+    City getCity(long id);
 
     City addCity(City city);
